Revert "test: move tests into functional folder"
This reverts commit 31cc389b8e1a07f9ba479e50bcfc5eb68d549691.
