+++ test/functional/query-builder/soft-delete/entity/User.ts
@@ -3,13 +3,7 @@ import { PrimaryGeneratedColumn } from "../../../../../src/decorator/columns/Pri
 import { DeleteDateColumn } from "../../../../../src/decorator/columns/DeleteDateColumn"
 import { Column } from "../../../../../src/decorator/columns/Column"
 import { Photo } from "./Photo"
-import {
-    JoinColumn,
-    OneToOne,
-    CreateDateColumn,
-    UpdateDateColumn,
-    VersionColumn,
-} from "../../../../../src"
+import { JoinColumn, OneToOne } from "../../../../../src"
 
 @Entity()
 export class User {
@@ -19,24 +13,12 @@ export class User {
     @Column()
     name: string
 
-    @Column({ nullable: true })
-    company: string
-
     @Column()
     likesCount: number = 0
 
     @OneToOne(() => Photo)
     @JoinColumn()
     picture: Photo
-
-    @CreateDateColumn()
-    createdAt: Date
-
-    @UpdateDateColumn()
-    updatedAt: Date
-
-    @VersionColumn()
-    version: number
 
     @DeleteDateColumn()
     deletedAt: Date
