+++ test/functional/query-builder/soft-delete/query-builder-soft-delete.test.ts
@@ -13,7 +13,6 @@ import { MissingDeleteDateColumnError } from "../../../../src/error/MissingDelet
 import { UserWithoutDeleteDate } from "./entity/UserWithoutDeleteDate"
 import { Photo } from "./entity/Photo"
 import { DriverUtils } from "../../../../src/driver/DriverUtils"
-import { scheduler } from "node:timers/promises"
 
 describe("query builder > soft-delete", () => {
     let connections: DataSource[]
@@ -320,172 +319,6 @@ describe("query builder > soft-delete", () => {
                     null,
                 )
                 expect(usersWithSoftDelete[1].picture.deletedAt).to.equal(null)
-            }),
-        ))
-
-    it("should only soft delete rows that are not soft deleted previously", () =>
-        Promise.all(
-            connections.map(async (connection) => {
-                const manager = connection.manager
-
-                // create test users
-                const batch1UsersData = [
-                    { name: "Hassan", company: "test1" },
-                    { name: "Nav", company: "test1" },
-                ]
-                const batch1Users = batch1UsersData.map(({ name, company }) => {
-                    const user = new User()
-                    user.name = name
-                    user.company = company
-                    return user
-                })
-                await manager.save(batch1Users)
-
-                // soft delete users with company test1
-                const del1 = await manager.softDelete(User, {
-                    company: "test1",
-                })
-                expect(del1.affected).to.be.eql(2)
-
-                // create more users with the same company test1
-                const batch2UsersData = [
-                    { name: "Omer", company: "test1" },
-                    { name: "Daniyal", company: "test1" },
-                    { name: "Salman", company: "test1" },
-                    { name: "Shahzaib", company: "test1" },
-                ]
-                const batch2Users = batch2UsersData.map(({ name, company }) => {
-                    const user = new User()
-                    user.name = name
-                    user.company = company
-                    return user
-                })
-                await manager.save(batch2Users)
-
-                // soft delete users again with company test1
-                const del2 = await manager.softDelete(User, {
-                    company: "test1",
-                })
-                // now affected rows should be equal to 4 and not 6, since 2 were already soft deleted before
-                expect(del2.affected).to.be.eql(batch2Users.length)
-            }),
-        ))
-
-    it("should correctly handle OR conditions in softDelete and not update already deleted rows", () =>
-        Promise.all(
-            connections.map(async (connection) => {
-                const manager = connection.manager
-
-                // create test users
-                const batch1UsersData = [
-                    { name: "User 10", company: "comp1" },
-                    { name: "User 11", company: "comp1" },
-                ]
-                const batch1Users = batch1UsersData.map((data) => {
-                    const user = new User()
-                    user.name = data.name
-                    user.company = data.company
-                    return user
-                })
-                await manager.save(batch1Users)
-
-                const ids1 = batch1Users.map((u) => u.id)
-
-                // soft delete users with ID 10 OR 11
-                const del1 = await manager.softDelete(
-                    User,
-                    ids1.map((id) => ({ id })),
-                )
-                expect(del1.affected).to.be.eql(2)
-
-                // create more users
-                const batch2UsersData = [
-                    { name: "User 12", company: "comp1" },
-                    { name: "User 13", company: "comp1" },
-                ]
-                const batch2Users = batch2UsersData.map((data) => {
-                    const user = new User()
-                    user.name = data.name
-                    user.company = data.company
-                    return user
-                })
-                await manager.save(batch2Users)
-
-                const ids2 = batch2Users.map((u) => u.id)
-
-                // soft delete users again with ID 10 OR 11 OR 12 OR 13
-                const del2 = await manager.softDelete(
-                    User,
-                    [...ids1, ...ids2].map((id) => ({ id })),
-                )
-
-                // now affected rows should be equal to 2 (batch2) and not 4, since 2 were already soft deleted before
-                expect(del2.affected).to.be.eql(batch2Users.length)
-
-                const softDeletedUsers = await manager.find(User, {
-                    where: [...ids1, ...ids2].map((id) => ({ id })),
-                    withDeleted: true,
-                })
-                expect(softDeletedUsers.length).to.be.eql(4)
-                softDeletedUsers.forEach((user) => {
-                    expect(user.deletedAt).to.be.instanceOf(Date)
-                })
-            }),
-        ))
-
-    it("should only restore rows that are soft deleted previously", () =>
-        Promise.all(
-            connections.map(async (connection) => {
-                const manager = connection.manager
-
-                // create test users
-                const user1 = new User()
-                user1.name = "User 1"
-                user1.company = "comp1"
-
-                const user2 = new User()
-                user2.name = "User 2"
-                user2.company = "comp1"
-
-                await manager.save([user1, user2])
-
-                // soft delete user 1
-                await manager.softDelete(User, { id: user1.id })
-
-                // user1 is now deleted, user2 is NOT.
-                // Both have company "comp1".
-
-                // Get current state of user 2
-                const user2Before = await manager.findOneBy(User, {
-                    id: user2.id,
-                })
-                expect(user2Before).to.exist
-                const updatedAtBefore = user2Before!.updatedAt
-                const versionBefore = user2Before!.version
-
-                // wait a bit to ensure CURRENT_TIMESTAMP will be different if updated
-                await scheduler.wait(1000)
-
-                // restore users with company comp1
-                const restoreResult = await manager.restore(User, {
-                    company: "comp1",
-                })
-
-                // only user1 should be restored
-                expect(restoreResult.affected).to.be.eql(1)
-
-                const user1After = await manager.findOneBy(User, {
-                    id: user1.id,
-                })
-                expect(user1After!.deletedAt).to.be.null
-
-                const user2After = await manager.findOneBy(User, {
-                    id: user2.id,
-                })
-                expect(user2After!.deletedAt).to.be.null
-                // updateDate and version should NOT have changed for user 2
-                expect(user2After!.updatedAt).to.be.eql(updatedAtBefore)
-                expect(user2After!.version).to.be.eql(versionBefore)
             }),
         ))
 })
